lib/algebra: the default {BCD} precision type is now {size_t} so it can be compared to {scale} without warnings

+++ lib/algebra/BCD.h
@@ -14,7 +14,7 @@
 
 namespace pyre {
     namespace algebra {
-        template <size_t scale, typename precision_t=long> class BCD;
+        template <size_t scale, typename precision_t=size_t> class BCD;
     }
 }
